Stone-Paper-Scissor-Game
Gameplay and Rules

Stone Paper Scissors is a hand game played between two people where each player simultaneously forms one of three shapes with their hand: a closed fist (representing a stone), an open hand (representing paper), or a hand with the index and middle fingers extended (representing scissors).

here in this the player has to pick one of the options by clicking on the respective object. Upon clicking, the computer makes its choice, and the result of the round can be seen immediately on screen.

The winner of the game is determined by a set of rules: rock crushes scissors, scissors cuts paper, and paper covers rock. If both players choose the same shape, the game is tied and played again.

and if player (user) will win then on clicking next button user redirect to the Hurray Page